Moved the functions starting with suggest to a separate file
